retriever/requirements.txt
fastapi
uvicorn
requests
pydantic
opensearch-py
pymilvus
python-dotenv
